chore: move mail utilities to a dedicated dir
