+++ backend/geonature/utils/mail/base_mail.py
